tree2: Move SchemaBuilder out of typed-schema (#17351)
## Description

Move SchemaBuilder out of typed-schema.
This will help with creating a new schema building API which can be
migrated to from the current one, as well as making schemaBuilder able
to link to editable tree for the canonical docs on the various field and
node kinds without that introducing a cycle.

Also renames SchemaCollection's file to better align with its current
content and our current naming.
